# drivers: ethernet: eth_stm32 add tx async mode