Enable TX transmission asynchronous mode to enhance performance
    
    Tested on the nucleo_h743zi board using samples/net/zperf
    
    west build -b nucleo_h743zi/stm32h743xx samples/net/zperf/ --pristine
    
    before:
    udp upload:   73.5Mbps/s
    tcp upload:   71.3Mbps/s
    udp download: 93.9Mbps/s
    tcp download: 70.5Mbps/s
    
    after:
    udp upload:   92.2Mbps/s
    tcp upload:   78.1Mbps/s
    udp download: 93.7Mbps/s
    tcp download: 75.2Mbps/s