@Freey0 2025-08-01
After adding the configuration, the performance has approached the level of other RTOS.

```
# For TCP
CONFIG_PICOLIBC_USE_MODULE=y
CONFIG_ETH_STM32_HW_CHECKSUM=y

#For UDP
CONFIG_NET_ZPERF_MAX_PACKET_SIZE=1470
```
udp upload:      94.1Mbps/s
tcp upload:        92.9Mbps/s
udp download:  93.7Mbps/s
tcp download:   94.5Mbps/s