@erwango 2025-10-15
> @erwango Could you help conduct comprehensive testing on the farm? Thank you very much.

@juickar Can you take this in charge ? This being said, this should only be considered as a simple non regression.
Tests performed in CI are pretty basic, so manual extra testing would be required.